backend/requirements.txt
fastapi
uvicorn
torchxrayvision
torch
torchvision
Pillow
numpy
scikit-image
python-multipart
